github-alerts/Google_Cloud_Function/requirements.txt
functions-framework==3.*
flask==2.1.0
slackclient==2.9.4